aop: make notes about best-practices in aop
